Add files via upload
This is a top trumps python project I am working on. The project uses an API to find pokemon. The game generates two random pokemons, one for the user and one for the opponent. The programme randomly selects a characteristic to compare, the user with the highest characteristic wins. The game runs four times and there is an overall winner. I am struggling with the part of the code that states the overall winner of the game Any help on this part would be appreciated.