Recipe proejct
This project was one completed in my Code First Girls career switcher course. The task was to create an app that pulled recipes from an API depending on the ingredient the user inputted. The programme was also coded to only show recipes that were vegetarian. The code also saves the results as a csv once it has been run.